Add IngestScheduler
For all your ingest scheduling needs
